+++ src/types/formio-metadata.ts
@@ -1,0 +1,10 @@
+export default interface FormioMetadata {
+  timezone?: string;
+  offset?: number;
+  origin?: string;
+  referrer?: string;
+  browserName?: string;
+  userAgent?: string;
+  pathName?: string;
+  onLine?: boolean;
+}