+++ src/types/formio-submission.ts
@@ -1,8 +1,11 @@
+import FormioMetadata from './formio-metadata';
+
 export default interface FormioSubmission<T = any, stateType= FormioSubmissionState > {
   _id?: string;
   created?: string;
   data?: T;
   form?: string;
+  metadata?: FormioMetadata;
   modified?: string;
   owner?: string;
   project?: string;