+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
+## [Unreleased]
+### Added
+ - ``FormioMetadata`` interface
 
 ## 4.8.0-rc.1
 ### Changed